+++ ReleaseNotes.md
@@ -7,6 +7,7 @@
 
 ### Bug Fixes
 - cta/CTA#259 - cta-rmcd should not exit if /dev/sg0 is missing
+- cta/CTA#345 - Remove some objectstore dependencies outside scheduler
 
 ### Continuous Integration
 - cta/CTA#262 - Manage CI repositories with `cta-release` code